Moved serialization classes to parent package due to needed non-public methods.
